requirements.txt
pytest~=8.2.2
selene~=2.0.0rc9
allure-pytest==2.13.5